Branching

Search

Split

Branching

The "branch" part of the Branch and Bound paradigm for verification algorithms. The branching folder contains algorithms for dividing the input set into searchable smaller sets, which we call "branches."

The search.jl module includes algorithms to iterate over all branches, such as BFS (Breadth-first Search) and DFS (Depth-first Search). The search.jl\search_branches function is of particular importance since it executes the verification procedure.

The split.jl module includes algorithms to split an unknown branch, such as bisect, sensitivity analysis, etc. The split.jl\split_branch function divides the unknown branches into smaller pieces and put them back to the branch bank for future verification. This is done so that we can get a more concrete answer by refining the problem in case the over-approximation introduced in the verification process prevents us from getting a firm result.

Search

ModelVerification.BFS — Type

BFS <: SearchMethod

Breadth-first Search (BFS)

Fields

max_iter :

batch_size :

ModelVerification.search_branches — Method

search_branches(search_method, split_method, prop_method, problem, model_info)

(Kai) Essentially THE FUNCTION that does the heavy lifting... Think of prop_method as the solver. Well, it is. This returns the result.

Split

ModelVerification.BaBSR — Type

BaBSR <: SplitMethod

Branch-and-Bound method for splitting.

ModelVerification.Bisect — Type

Bisect <: SplitMethod

ModelVerification.InputGradSplit — Type

InputGradSplit <: SplitMethod

ModelVerification.branching_scores_kfsb — Method

branching_scores_kfsb(model_info, batch_info, input)

"Kernel Function Split Branch"

ModelVerification.split_beta — Method

split_beta(S_dict, score, split_relu_node, i, split_neurons_index_in_node, j, input, output)

ModelVerification.split_branch — Method

split_branch(split_method::BaBSR, model::Chain, input::Tuple, output, model_info, batch_info)

ModelVerification.split_branch — Method

split_branch(split_method::Bisect, model::Chain, input::Hyperrectangle, output, model_info, batch_info)

ModelVerification.split_branch — Method

split_branch(split_method::Bisect, model::Chain, input::LazySet, output, model_info, batch_info)

ModelVerification.split_branch — Method

split_branch(split_method::Bisect, model::Chain, input::ImageStarBound, output, model_info, batch_info)

ModelVerification.split_branch — Method

split_branch(split_method::Bisect, model::Chain, input::ImageStarBound, output)

ModelVerification.split_branch — Method

split_branch(split_method::Bisect, model::Chain, input::ImageZonoBound, output, model_info, batch_info)

ModelVerification.split_branch — Method

split_branch(split_method::Bisect, model::Chain, input::ImageZonoBound, output)

ModelVerification.split_interval — Method

split_interval(dom, i)

Split a set into two at the given index.

Inputs:

dom::Hyperrectangle: the set to be split

i: the index to split at

Return:

(left, right)::Tuple{Hyperrectangle, Hyperrectangle}: two sets after split

ModelVerification.topk — Method

topk(score, k, model_info)

"Top Kernel"

ModelVerification.vecmask_convert_to_original_size — Method

vecmask_convert_to_original_size(index, original)

ModelVerification.vecsign_convert_to_original_size — Method

vecsign_convert_to_original_size(index, vector, original)