Network

Model

Network

Activation Functions

Helper Functions

Network

Model

ModelVerification.Model — Type

Model

Structure containing the information of the neural network to be verified.

Fields

start_nodes (Array{String, 1}): List of input layer nodes' names.

final_nodes (Array{String, 1}): List of output layer nodes' names.

all_nodes (Array{String, 1}): List of all the nodes's names.

node_layer (Dict): Dictionary of all the nodes. The key is the name of the node and the value is the operation performed at the node.

node_prevs (Dict): Dictionary of the nodes connected to the current node. The key is the name of the node and the value is the list of nodes.

node_nexts (Dict): Dictionary of the nodes connected from the current node. The key is the name of the node and the value is the list of nodes.

activation_nodes (Array{String, 1}): List of all the activation nodes' names.

activation_number (Int): Number of activation nodes (deprecated in the future).

Network

ModelVerification.Network — Type

Network

A vector of layers.

Fields

layers (Vector{Layer}): Layers of the network, including the output layer.

See also: Layer

ModelVerification.Layer — Type

Layer{F<:ActivationFunction, N<:Number}

Consists of weights and bias for linear mapping, and activation for nonlinear mapping.

Fields

weights::Matrix{N}

bias::Vector{N}

activation::F

See also: Network

Activation Functions

ModelVerification.ActivationFunction — Type

ActivationFunction

Function that calculates the output of the node. Supported activation functions are:

ReLU (ReLU)

Max (Max)

Identity (Id)

Sigmoid (Sigmoid)

Tanh (Tanh)

ModelVerification.GeneralAct — Type

GeneralAct <: ActivationFunction

Wrapper type for a general activation function.

Usage

act = GeneralAct(tanh)

act(0) == tanh(0)           # true
act(10.0) == tanh(10.0)     # true

act = GeneralAct(x->tanh.(x))

julia> act(-2:2)
5-element Array{Float64,1}:
 -0.9640275800758169
 -0.7615941559557649
  0.0
  0.7615941559557649
  0.9640275800758169

ModelVerification.Id — Type

Id <: ActivationFunction

Identity operator

(Id())(x) -> x

ModelVerification.Max — Type

Max <: ActivationFunction

(Max())(x) -> max(maximum(x), 0)

ModelVerification.PiecewiseLinear — Type

PiecewiseLinear <: ActivationFunction

Activation function that uses linear interpolation between supplied knots. An extrapolation condition can be set for values outside the set of knots. Default is Linear.

PiecewiseLinear(knots_x, knots_y, [extrapolation = Line()])

Usage

kx = [0.0, 1.2, 1.7, 3.1]
ky = [0.0, 0.5, 1.0, 1.5]
act = PiecewiseLinear(kx, ky)

act(first(kx)) == first(ky) == 0.0
act(last(kx))  == last(ky)  == 1.5

act(1.0)    # 0.4166666666666667
act(-102)   # -42.5

act = PiecewiseLinear(kx, ky, Flat())

act(-102)   # 0.0
act(Inf)    # 1.5

Extrapolations

Flat()

Line()

constant (supply a number as the argument)

Throw() (throws bounds error)

PiecewiseLinear uses Interpolations.jl.

ModelVerification.ReLU — Type

ReLU <: ActivationFunction

(ReLU())(x) -> max.(x, 0)

ModelVerification.Sigmoid — Type

Sigmoid <: ActivationFunction

(Sigmoid())(x) -> 1 ./ (1 .+ exp.(-x))

ModelVerification.Tanh — Type

Tanh <: ActivationFunction

(Tanh())(x) -> tanh.(x)

Helper Functions

ModelVerification.onnx_parse — Method

onnx_parse(onnx_model_path)

Creates the Model from the onnx_model_path. First, the computational graph of the ONNX model is created. Then, the Model is created using the information retrieved from the computational graph.

Arguments

onnx_model_path: String path to the ONNX model.

Returns

model_info (Model): Contains network information retrieved from the computational graph of the ONNX model.

ModelVerification.get_act — Method

get_act(l)

Returns the activation function of the node l if it exists.

Arguments

l: node

Returns

Activation function of the node if it exists.

Otherwise, return nothing.

ModelVerification.n_nodes — Method

n_nodes(L::Layer)

Returns the number of neurons in a layer.

ModelVerification.read_nnet — Method

read_nnet(fname::String; last_layer_activation = Id())

Read in neural net from a .nnet file and return Network struct. The .nnet format is borrowed from NNet. The format assumes all hidden layers have ReLU activation. Keyword argument last_layer_activation sets the activation of the last layer, and defaults to Id(), (i.e. a linear output layer).

Arguments

fname (String): String path to the .nnet file.

last_layer_activation: Keyword argument that sets the activation of the last layer which defaults to Id().

Returns

A vector of layers saved as Network.

ModelVerification.read_layer — Function

read_layer(output_dim::Int64, f::IOStream, act = ReLU())

Read in layer from .nnet file and return a Layer containing its weights & biases. Optional argument act sets the activation function for the layer.

Arguments

output_dim (Int64): Output dimension of the layer.

f (IOStream): IO stream of the .nnet file.

act: Optional argument specifying the activation function of the layer. Defaults to ReLU().

Returns

Layer containing the weights and bias values (and the activation function of the layer).

ModelVerification.to_comment — Method

to_comment(txt)

Prepend // to each line of a string.

ModelVerification.print_layer — Method

print_layer(file::IOStream, layer)

Print to file an object implementing weights(layer) and bias(layer).

Arguments

file (IOStream): IO stream of the target .nnet file.

layer: Layer to be transcribed to file.

ModelVerification.print_header — Method

print_header(file::IOStream, network; header_text="")

The NNet format has a particular header containing information about the network size and training data. print_header does not take training-related information into account (subject to change).

Arguments

file (IOStream): IO stream of the target .nnet file.

network: Network to be transcribed to file.

header_text: Optional header text that comes before the network information. Defaults to an empty string.

ModelVerification.write_nnet — Method

write_nnet(filename, network; header_text)

Write network to filename.nnet. Note: Does not perform safety checks on inputs, so use with caution.

Based on python code at https://github.com/sisl/NNet/blob/master/utils/writeNNet.py and follows .nnet format given here: https://github.com/sisl/NNet.

Arguments

outfile: String name of the .nnet file.

network: Network to be transcribed to outfile.nnet.

header_text: Optional header text that comes before the network information.