Problem Outline

Network

Safety Property

Solver

Problem Outline

Verification checks if the input-output relationships of a function, specifically deep neural networks in this case, hold. For an input specification imposed by a set $\mathcal{X}\subseteq \mathcal{D}_x$, we would like to check if the corresponding output of the function is contained in an output specification imposed by a set $\mathcal{Y}\subseteq \mathcal{D}_y$:

\[x\in\mathcal{X} \Longrightarrow y = f(x) \in \mathcal{Y}.\]

ModelVerification.AdversarialResult — Type

AdversarialResult(status, max_disturbance)

Like BasicResult, but also returns the maximum allowable disturbance in the input (if status = :violated).

ModelVerification.BasicResult — Type

BasicResult(status::Symbol)

Result type that captures whether the input-output constraint is satisfied. Possible status values:

:holds (io constraint is satisfied always)

:violated (io constraint is violated)

:unknown (could not be determined)

ModelVerification.CounterExampleResult — Type

CounterExampleResult(status, counter_example)

Like BasicResult, but also returns a counter_example if one is found (if status = :violated). The counter_example is a point in the input set that, after the NN, lies outside the output set.

ModelVerification.Problem — Type

Problem{P, Q}(network::Network, input::P, output::Q)

Problem definition for neural verification.

The verification problem consists of: for all points in the input set, the corresponding output of the network must belong to the output set.

Fields

network : Network that can be constructed either using the path to an onnx model or a Flux.Chain structure.

input : input specification defined using a LazySet.

output : output specification defined using a LazySet.

ModelVerification.ReachabilityResult — Type

ReachabilityResult(status, reachable)

Like BasicResult, but also returns the output reachable set given the input constraint (if status = :violated).

ModelVerification.Result — Type

Result

Supertype of all result types.

See also: BasicResult, CounterExampleResult, AdversarialResult, ReachabilityResult

Network

Details on Network

Safety Property

Details on [Input-Output Specification](./safetyspec.md)_

Solver

Details on Solvers