Input-Output Specification

Safety Property

Geometric Representation

Hyperrectangle (Hyperrectangle)

HalfSpace (HalfSpace)

Halfspace-Polytope (HPolytope)

StarSet (Star)

Zonotope (Zonotope)

ImageStar

ImageZono

PolytopeComplement

References

Spec

Input-Output Specification

Safety Property

A safety property is essentially an input-output relationship for the model we want to verify. In general, the constraints for the input set $\mathcal{X}$ and the output set $\mathcal{Y}$ can have any geometry. For the sake of simplicity, ModelVerification.jl uses convex polytopes and the complement of a polytope to encode the input and output specifications. Specifically, our implementation utilizes the geometric definitions of LazySets, a Julia package for calculus with convex sets. The following section dives into the geometric representations ModelVerification.jl uses and the representations required for each solver.

Geometric Representation

Different solvers implemented in ModelVerification.jl require the input-output specification formulated with particular geometries. We report here a brief overview of the sets we use. For specifics, please read Algorithms for Verifying Deep Neural Networks by C. Liu, et al. and Sets in LazySets.jl.

HR = Hyperrectangle

HS = HalfSpace

HP = HPolytope

SS = StarSet

ZT = Zonotope

IS = ImageStar

IZ = ImageZono

PC = PolytopeComplement

Solver	Input Set	Output Set

Ai2	HR, ZT, HP, SS	HP (bounded)

CROWN	HR	HP (bounded)

$\alpha$-CROWN	HR	HP (bounded)

$\beta$-CROWN	HR	HP (bounded)

$\alpha$-$\beta$-CROWN	HR	HP (bounded)

Hyperrectangle (Hyperrectangle)

Corresponds to a high-dimensional rectangle, defined by

\[|x-c| \le r,\]

where $c\in\mathbb{R}^{k_0}$ is the center of the hyperrectangle and $r\in\mathbb{R}^{k_0}$ is the radius of the hyperrectangle.

HalfSpace (HalfSpace)

Represented by a single linear inequality constraint

\[c^\top x \le d,\]

where $c\in\mathbb{R}^{k_0}$ and $d\in\mathbb{R}$.

Halfspace-Polytope (HPolytope)

HPolytope uses a set of linear inequality constraints to represent a convex polytope, i.e., it is a bounded set defined using an intersection of half-spaces.

\[Cx \le d,\]

where $C\in\mathbb{R}^{k\times k_0}, d\in\mathbb{R}^k$ with $k$ representing the number of inequality constraints.

StarSet (Star)

Only convex star set is considered in this toolbox. A convex star set is an affine transformation of an arbitrary convex polytope,

\[x = c + \begin{bmatrix} r_1 & r_2 & \cdots & r_l \end{bmatrix} \alpha,\; C\alpha \le d,\]

where $c\in\mathbb{R}^{k_0}$ is the center of the star set, $r_i\in\mathbb{R}^{k_0},\; i\in\{1,\dots,l\}$ are generators of the star set, $C\in\mathbb{R}^{k\times l}$, $d\in\mathbb{R}^{k}$, $\alpha\in\mathbb{R}^l$ is the free parameter that belongs to a unit hypercube, and $k$ is the number of inequality constraints on $\alpha$. $l$ is the degree of freedom of the star set.

Zonotope (Zonotope)

Zonotope is basically as star set in which all predicate variables are in the range of $[-1, 1]$. Zonotope represents polytopes that can be written as affine transformations of a unit hypercube, defined as

\[x = c + \begin{bmatrix} r_1 & r_2 & \cdots & r_l \end{bmatrix} \alpha,\; |\alpha| \le 1,\]

where $c\in\mathbb{R}^{k_0}$ is the center of the zonotope, $r_i\in\mathbb{R}^{k_0},\; i\in\{1,\dots,l\}$ are generators of the zonotope, and $\alpha\in\mathbb{R}^l$ is the free parameter that belongs to a unit hypercube. $l$ is the degree of freedom of the zonotope.

ImageStar

ImageStar is an extension of the star set where the center and generators are images with multiple channels.

\[x = c + \begin{bmatrix} r_1 & r_2 & \cdots & r_l \end{bmatrix} \alpha,\; C\alpha \le d,\]

where $c\in\mathbb{R}^{h\times w \times k_0}$ is the center image, $r_i\in\mathbb{R}^{h \times w \times k_0},\; i\in\{1,\dots,l\}$ are the generator iamges, $C\in\mathbb{R}^{k\times l}$, $d\in\mathbb{R}^{k}$, and $h,w,k$ are the height, width, and number of channels (input dimension) of the images respectively. $\alpha\in\mathbb{R}^l$ is the free parameter that belongs to a unit hypercube, and $k$ is the number of inequality constraints on $\alpha$. $l$ is the degree of freedom of the star set.

ImageZono

PolytopeComplement

References

[1] C. Liu, T. Arnon, C. Lazarus, C. Strong, C. Barret, and M. J. Kochenderfer, "Algorithms for Verifying Deep Neural Networks," in Foundations and Trends in Optimization, 2021.

[2] T. Gehr, M. Mirman, D. Drashsler-Cohen, P. Tsankov, S. Chaudhuri, and M. Vechev, "Ai2: Safety and Robustness Certification of Neural Networks with Abstract Interpretation," in 2018 IEEE Symposium on Security and Privacy (SP), 2018.

[3] M. Forets and C. Schilling, "LazySets.jl: Scalable Symbolic-Numeric Set Computations," in Proceeds of the JuliaCon Conferences, 2021.

[4] HD. Tran, S. Bak, W. Xiang, and T.T. Johnson, "Verification of Deep Convolutional Neural Networks Using ImageStars," in Computer Aided Verification (CAV), 2020.

Spec

ModelVerification.InputSpec — Type

InputSpec

Input specification can be of any type supported by LazySet or ImageConvexHull.

ModelVerification.OutputSpec — Type

OutputSpec

Output specification can be of any type supported by LazySet or LinearSpec.

ModelVerification.ImageConvexHull — Type

ImageConvexHull <: Spec

Convex hull for images used to specify safety property for images. It is the smallest convex polytope that contains all the images given in the imgs array.

Fields

imgs (AbstractArray): list of images in AbstractArray. Image is represented as a matrix of h x w x c.

ModelVerification.LinearSpec — Type

LinearSpec <: Spec

Safety specification defined as the set $x: x = A x - b ≤ 0$.

Fields

A (AbstractArray{Float64, 3}): normal dierction of size spec_dim x out_dim x batch_size.

b (AbstractArray{Float64, 2}): constraints of size spec_dim x batch_size.

is_complement (Bool): boolean flag for whether this specification is a complement or not.

ModelVerification.Spec — Type

Spec

Abstract super-type for input-output specifications.

ModelVerification.classification_spec — Method

classification_spec(n, target)

ModelVerification.get_linear_spec — Method

get_linear_spec(batch_out_set::AbstractVector)

Retrieves the linear specifications of the batch of output sets and returns a LinearSpec structure.

Fields:

batch_out_set (AbstractVector): batch of output sets.

Returns:

LinearSpec of the batch of output sets.

ModelVerification.get_size — Method

get_size(input_spec::ImageConvexHull)

Given an ImageConvexHull, it determines the size.

ModelVerification.get_size — Method

get_size(input_spec::LazySet)

Given a LazySet, it determines the size.