Solvers

Ai2

ImageStar

Crown

$\alpha$-Crown

$\beta$-Crown

Solvers

ModelVerification.BackwardProp — Type

BackwardProp <: PropMethod

ModelVerification.BatchBackwardProp — Type

BatchBackwardProp <: BackwardProp

ModelVerification.BatchForwardProp — Type

BatchForwardProp <: ForwardProp

ModelVerification.Bound — Type

Bound

ModelVerification.ForwardProp — Type

ForwardProp <: PropMethod

ModelVerification.SequentialBackwardProp — Type

SequentialBackwardProp <: ForwardProp

ModelVerification.SequentialForwardProp — Type

SequentialForwardProp <: ForwardProp

ModelVerification.check_inclusion — Method

check_inclusion(prop_method::ForwardProp, model, batch_input::AbstractArray, batch_reach::AbstractArray, batch_output::AbstractArray)

ModelVerification.init_batch_bound — Method

init_batch_bound(prop_method::BackwardProp, batch_input, batch_output)

ModelVerification.init_batch_bound — Method

init_batch_bound(prop_method::ForwardProp, batch_input, batch_output)

ModelVerification.init_bound — Method

init_bound(prop_method::BackwardProp, output)

ModelVerification.init_bound — Method

init_bound(prop_method::ForwardProp, input)

ModelVerification.init_propagation — Method

init_propagation(prop_method::BackwardProp, batch_input, batch_output, model_info)

ModelVerification.init_propagation — Method

init_propagation(prop_method::ForwardProp, batch_input, batch_output, model_info)

ModelVerification.init_propagation — Method

function initbound(propmethod::ForwardProp, input) return input end

function initbound(propmethod::BackwardProp, output) return output end

function processbound(propmethod::PropMethod, batchbound, batchoutspec, modelinfo, batchinfo) return batchbound, batch_info end

ModelVerification.prepare_method — Method

prepare_method(prop_method::PropMethod, batch_input::AbstractVector, batch_output::AbstractVector, model_info)

ModelVerification.process_bound — Method

process_bound(prop_method::PropMethod, batch_bound, batch_out_spec, model_info, batch_info)

Ai2

ModelVerification.Ai2 — Type

Ai2{T}

Ai2 performs over-approximated reachability analysis to compute the over- approximated output reachable set for a network. T can be Hyperrectangle, Zonotope, Star, or HPolytope, and determines the amount of over- approximation (and hence also performance tradeoff). The original implementation (from [1]) uses Zonotopes, so we consider this the "benchmark" case. The HPolytope case is more precise, but not scalable, and the opposite is true of the Hyperrectangle case. Zonotope and Star are more scalable and precise.

Note that initializing Ai2() defaults to Ai2{Zonotope}. The following aliases also exist for convenience:

const Ai2h = Ai2{HPolytope}
const Ai2z = Ai2{Zonotope}
const Ai2s = Ai2{Star}
const Box = Ai2{Hyperrectangle}

Problem requirement

Network: any depth, ReLU activation (more activations to be supported in the future)

Input: AbstractPolytope

Output: AbstractPolytope

Return

ReachabilityResult

Method

Reachability analysis using split and join.

Property

Sound but not complete.

Note

Efficient over-approximation of intersections and unions involving zonotopes relies on Theorem 3.1 of [2].

Reference

[1] T. Gehr, M. Mirman, D. Drashsler-Cohen, P. Tsankov, S. Chaudhuri, and M. Vechev, "Ai2: Safety and Robustness Certification of Neural Networks with Abstract Interpretation," in 2018 IEEE Symposium on Security and Privacy (SP),

[2] Singh, G., Gehr, T., Mirman, M., Püschel, M., & Vechev, M. (2018). Fast and effective robustness certification. In Advances in Neural Information Processing Systems (pp. 10802-10813).

ModelVerification.StarSet — Type

StarSet

Covers all Ai2 variations: Ai2h, Ai2z, Ai2s, Box.

Fields

pre_bound_method:

ModelVerification.check_inclusion — Method

check_inclusion(prop_method::ForwardProp, model, input::LazySet, 
reach::LazySet, output::Complement)

Determines whether the reachable set, R(input, model), is within the valid output specified by a LazySet. This function achieves this by checking if the box approximation (overapproximation with hyperrectangle) of the reach set is disjoint with the unsafe_output.

Fields

prop_method :

model : deep neural network model that is to be verified.

input : input specification represented using a Lazyset.

reach : reachable set resulting from the propagation of input through the model.

output : set of valid outputs represented with a complement set.

Returns

ReachabilityResult(:holds, [reach]) if box_reach is disjoint with the complement of the output.

CounterExampleResult(:violated, x) if the center of the input set results in a state that belongs to the unsafe_output.

CounterExampleResult(:unknown) if either the two cases above are true.

ModelVerification.check_inclusion — Method

check_inclusion(prop_method::ForwardProp, model, input::LazySet, 
reach::LazySet, output::LazySet)

Determines whether the reachable set, R(input, model), is within the valid output specified by a LazySet. This function achieves this by directly checking if the reachable set reach is a subset of the set of valid outputs output. If not, it attempts to find a counterexample and returns the appropriate Result.

Fields

prop_method : (need to be deprecated).

model : deep neural network model that is to be verified.

input : (need to be deprecated).

reach : reachable set resulting from the propagation of input through the model.

output : set of valid outputs represented with a LazySet.

Returns

ReachabilityResult(:holds, [reach]) if reach is a subset of output.

CounterExampleResult(:unknown) if reach is not a subset of output, but cannot find a counterexample.

CounterExampleResult(:violated, x) if reach is not a subset of output, and there is a counterexample.

ModelVerification.compute_bound — Method

compute_bound(bound::Star)

Computes the lower- and upper-bounds of a star set. This function is used when propagating through the layers of network. It overapproximates the given star set with a hyperrectangle.

Fields

bound : star of which the bounds need to be computed

Returns

Lower- and upper-bounds of the overapproximated hyperrectangle.

ModelVerification.compute_bound — Method

compute_bound(bound::Zonotope)

Computes the lower- and upper-bounds of a zonotope. This function is used when propagating through the layers of network. Radius is the sum of absolute value of the generators of the given zonotope.

Fields

bound : zonotope of which the bounds need to be computed

Returns

Lower- and upper-bounds of the Zonotope.

ModelVerification.init_bound — Method

init_bound(prop_method::StarSet, input::Hyperrectangle)

Given a hyperrectangle as input, this function returns a star set that encompasses the hyperrectangle. This helps a more precise computation of bounds.

Fields

prop_method : (not used –> need to be deprecated)

input : hyperrectangle to be converted into a star set

Returns

Star set that encompasses the given hyperrectangle.

ModelVerification.prepare_method — Method

prepare_method(prop_method::StarSet, batch_input::AbstractVector, 
batch_output::AbstractVector, model_info)

Fields

prop_method : propagation method of type StarSet.

batch_input :

batch_output :

model_info :

ImageStar

ModelVerification.ImageStar — Type

ImageStar <: SequentialForwardProp

ImageStar is a verification approach that can verify the robustness of CNN. It is defined as a set representation in literature, but this toolbox uses the term as the verification method itself that uses the ImageStar set. ImageStarBound is used to represent the bounded set.

Fields

pre_bound_method :

Reference

[1] HD. Tran, S. Bak, W. Xiang, and T.T. Johnson, "Verification of Deep Convolutional Neural Networks Using ImageStars," in Computer Aided Verification (CAV), 2020.

ModelVerification.ImageStarBound — Type

ImageStarBound{T<:Real} <: Bound

ModelVerification.assert_zono_star — Method

assert_zono_star(bound::ImageStarBound)

ModelVerification.check_inclusion — Method

check_inclusion(prop_method::ImageStar, model, input::ImageStarBound, reach::LazySet, output::LazySet)

ModelVerification.compute_bound — Method

compute_bound(bound::ImageStarBound)

ModelVerification.init_bound — Method

init_bound(prop_method::ImageStar, ch::ImageConvexHull)

Assume batch_input[1] is a list of vertex images. Return a zonotope.

Outputs:

ImageStarBound

ModelVerification.prepare_problem — Method

prepare_problem(search_method, split_method, prop_method, problem)

Arguments

search_method::SearchMethod:

split_method::SplitMethod:

prop_method::ImageStar:

problem::Problem:

ModelVerification.ImageZono — Type

ImageZono <: SequentialForwardProp

ModelVerification.ImageZonoBound — Type

ImageZonoBound{T<:Real} <: Bound

ModelVerification.check_inclusion — Method

check_inclusion(prop_method::ImageZono, model, input::ImageZonoBound, reach::LazySet, output::LazySet)

ModelVerification.compute_bound — Method

compute_bound(bound::ImageZonoBound)

ModelVerification.init_bound — Method

init_bound(prop_method::ImageZono, ch::ImageConvexHull)

ModelVerification.init_bound — Method

init_bound(prop_method::ImageZono, bound::ImageStarBound)

ModelVerification.prepare_problem — Method

prepare_problem(search_method::SearchMethod, split_method::SplitMethod, prop_method::ImageZono, problem::Problem)

Crown

$\alpha$-Crown

$\beta$-Crown