Network

Network

Activation Functions

Helper Functions

Network

Network

ModelVerification.Layer — Type

Layer{F, N}

Consists of weights and bias for linear mapping, and activation for nonlinear mapping.

Fields

weights::Matrix{N}

bias::Vector{N}

activation::F

See also: Network

ModelVerification.Network — Type

A Vector of layers.

Network([layer1, layer2, layer3, ...])

See also: Layer

ModelVerification.n_nodes — Method

n_nodes(L::Layer)

Returns the number of neurons in a layer.

Activation Functions

ModelVerification.ActivationFunction — Type

ActivationFunction

Function that calculates the output of the node. Supported activation functions are:

ReLU (ReLU)

Max (Max)

Identity (Id)

Sigmoid (Sigmoid)

Tanh (Tanh)

ModelVerification.GeneralAct — Type

GeneralAct <: ActivationFunction

Wrapper type for a general activation function.

Usage

act = GeneralAct(tanh)

act(0) == tanh(0)           # true
act(10.0) == tanh(10.0)     # true

act = GeneralAct(x->tanh.(x))

julia> act(-2:2)
5-element Array{Float64,1}:
 -0.9640275800758169
 -0.7615941559557649
  0.0
  0.7615941559557649
  0.9640275800758169

ModelVerification.Id — Type

Id <: ActivationFunction

Identity operator

(Id())(x) -> x

ModelVerification.Max — Type

Max <: ActivationFunction

(Max())(x) -> max(maximum(x), 0)

ModelVerification.PiecewiseLinear — Type

PiecewiseLinear <: ActivationFunction

Activation function that uses linear interpolation between supplied knots. An extrapolation condition can be set for values outside the set of knots. Default is Linear.

PiecewiseLinear(knots_x, knots_y, [extrapolation = Line()])

Usage

kx = [0.0, 1.2, 1.7, 3.1]
ky = [0.0, 0.5, 1.0, 1.5]
act = PiecewiseLinear(kx, ky)

act(first(kx)) == first(ky) == 0.0
act(last(kx))  == last(ky)  == 1.5

act(1.0)    # 0.4166666666666667
act(-102)   # -42.5

act = PiecewiseLinear(kx, ky, Flat())

act(-102)   # 0.0
act(Inf)    # 1.5

Extrapolations

Flat()

Line()

constant (supply a number as the argument)

Throw() (throws bounds error)

PiecewiseLinear uses Interpolations.jl.

ModelVerification.ReLU — Type

ReLU <: ActivationFunction

(ReLU())(x) -> max.(x, 0)

ModelVerification.Sigmoid — Type

Sigmoid <: ActivationFunction

(Sigmoid())(x) -> 1 ./ (1 .+ exp.(-x))

ModelVerification.Tanh — Type

Tanh <: ActivationFunction

(Tanh())(x) -> tanh.(x)

Helper Functions

ModelVerification.act_gradient — Method

act_gradient(act, z_hat::Vector{N}) where N

Compute the gradient of an activation function at point z_hat. Currently only supports ReLU and Id.

ModelVerification.act_gradient_bounds — Method

act_gradient_bounds(nnet::Network, input::AbstractPolytope)

Compute the bounds on the gradient of all activation functions given an input set. Currently only support ReLU. Return:

LΛ, UΛ::NTuple{2, Vector{BitVector}}: lower and upper bounds on activation

ModelVerification.affine_map — Method

affine_map(layer, x)

Compute W*x ⊕ b for a vector or LazySet x

ModelVerification.affine_map — Method

affine_map(layer, x)

Compute W*x ⊕ b for a vector or LazySet x

ModelVerification.approximate_act_map — Method

approximateactmap(layer, input::Hyperrectangle)

Returns a Hyperrectangle overapproximation of the activation map of the input. actmust be monotonic.

ModelVerification.approximate_affine_map — Method

approximateaffinemap(layer, input::Hyperrectangle)

Returns a Hyperrectangle overapproximation of the affine map of the input.

ModelVerification.broadcast_mid_dim — Method

broadcast_mid_dim(m::AbstractArray{2}, target::AbstractArray{T,3})

Given a target tensor of the shape AxBxC, broadcast the 2D mask of the shape AxC to AxBxC.

Outputs:

m broadcasted.

ModelVerification.compute_output — Method

compute_output(nnet::Network, input::Vector{Float64})

Propagate a given vector through a nnet and compute the output.

ModelVerification.get_activation — Method

get_activation(L::Layer{ReLU}, bounds::Hyperrectangle)

Given a layer, find the activation pattern of all neurons in the layer given the node-wise bounds. Assume ReLU. Assume bounds is the pre-activation bounds for each ReLU in the layer. return Vector{Vector{Int64}}.

1: activated

0: undetermined

-1: not activated

ModelVerification.get_activation — Method

get_activation(L, x::Vector)

Finds the activation pattern of a vector x subject to the activation function given by the layer L. Returns a Vector{Bool} where true denotes the node is "active". In the sense of ReLU, this would be x[i] >= 0.

ModelVerification.get_activation — Method

get_activation(nnet::Network, input::Hyperrectangle)

Given a network, find the activation pattern of all neurons for a given input set. Assume ReLU. return Vector{Vector{Int64}}.

1: activated

0: undetermined

-1: not activated

ModelVerification.get_activation — Method

get_activation(nnet::Network, x::Vector)

Given a network, find the activation pattern of all neurons at a given point x. Returns Vector{Vector{Bool}}. Each Vector{Bool} refers to the activation pattern of a particular layer.

ModelVerification.get_activation — Method

get_activation(nnet::Network, bounds::Vector{Hyperrectangle})

Given a network, find the activation pattern of all neurons given the node-wise bounds. Assume ReLU. Assume pre-activation bounds where the bounds on the input are given by the first hyperrectangle, the first hidden layer by the second hyperrectangle, and so on. return Vector{Vector{Int64}}.

1: activated

0: undetermined

-1: not activated

ModelVerification.get_bounds — Method

get_bounds(problem::Problem)
get_bounds(nnet::Network, input::Hyperrectangle, [true])

Computes node-wise bounds given a input set. The optional last argument determines whether the bounds are pre- or post-activation.

Return:

Vector{Hyperrectangle}: bounds for all nodes. bounds[1] is the input set.

ModelVerification.get_gradient — Method

get_gradient(nnet::Network, x::Vector)

Given a network, find the gradient at the input x

ModelVerification.get_gradient_bounds — Method

get_gradient_bounds(nnet::Network, LΛ::Vector{AbstractVector}, UΛ::Vector{AbstractVector})
get_gradient_bounds(nnet::Network, input::AbstractPolytope)

Get lower and upper bounds on network gradient for given gradient bounds on activations, or given an input set. Return:

(LG, UG)::NTuple{2, Matrix{Float64} lower and upper bounds.

ModelVerification.interval_map — Method

interval_map(W::Matrix, l::AbstractVecOrMat, u::AbstractVecOrMat)

Simple linear mapping on intervals. L, U := ([W]₊*l + [W]₋*u), ([W]₊*u + [W]₋*l)

Outputs:

(lbound, ubound) (after the mapping)

ModelVerification.print_header — Method

print_header(file::IOStream, network[; header_text])

The NNet format has a particular header containing information about the network size and training data. print_header does not take training-related information into account (subject to change).

ModelVerification.print_layer — Method

print_layer(file::IOStream, layer)

Print to file an object implementing weights(layer) and bias(layer)

ModelVerification.read_layer — Function

read_layer(output_dim::Int, f::IOStream, [act = ReLU()])

Read in layer from nnet file and return a Layer containing its weights/biases. Optional argument act sets the activation function for the layer.

ModelVerification.read_nnet — Method

read_nnet(fname::String; last_layer_activation = Id())

Read in neural net from a .nnet file and return Network struct. The .nnet format is borrowed from NNet. The format assumes all hidden layers have ReLU activation. Keyword argument last_layer_activation sets the activation of the last layer, and defaults to Id(), (i.e. a linear output layer).

ModelVerification.relaxed_relu_gradient — Method

relaxed_relu_gradient(l::Real, u::Real)

Return the slope of a ReLU activation based on its lower and upper bounds

Returns 0 if u<0, 1 if l>0, u/(u-l) otherwise

ModelVerification.to_comment — Method

Prepend // to each line of a string.

ModelVerification.write_nnet — Method

write_nnet(filename, network[; header_text])

Write network to filename.nnet. Note: Does not perform safety checks on inputs, so use with caution.

Based on python code at https://github.com/sisl/NNet/blob/master/utils/writeNNet.py and follows .nnet format given here: https://github.com/sisl/NNet.