Propagation

Functions for propagating the bound through the model (from start nodes to the end nodes) for a given branch. For a forward propagation method (ForwardProp), the start nodes are the input nodes of the computational graph and the end nodes are the output nodes. For a backward propagation method (BackwardProp), the start nodes are the output nodes and the end nodes are the input nodes. We use BFS (Breadth-first Search) to iterate through the computational graph and propagates the bounds from nodes to nodes.

The propagate\propagate.jl module defines algorithms for propagating bounds from input to output, for both forward propagation and backward propagation.

The propagate\operators folder contains specific propagation algorithms for different operators, such as ReLU, Dense, Identity, Convolution, Bivariate, etc.

ModelVerification.PropMethod — Type

PropMethod

ModelVerification.backward_layer — Method

backward_layer(prop_method, layer, batch_bound)

ModelVerification.enqueue_nodes! — Method

enqueue_nodes!(prop_method::BackwardProp, queue, model_info)

Inserts the final nodes into the given queue.

ModelVerification.enqueue_nodes! — Method

enqueue_nodes!(prop_method::ForwardProp, queue, model_info)

Inserts the nodes connected from the starting node into the given queue.

ModelVerification.forward_act_batch — Method

forward_act_batch(prop_method::ForwardProp, σ, batch_reach::AbstractArray, batch_info::AbstractArray)

ModelVerification.forward_layer — Method

forward_layer(prop_method, layer, batch_bound, batch_info)

ModelVerification.forward_linear_batch — Method

forward_linear_batch(prop_method::ForwardProp, layer, batch_reach::AbstractArray, batch_info::AbstractArray)

ModelVerification.forward_skip_batch — Method

forward_skip_batch(prop_method::ForwardProp, layer, batch_reach1::AbstractArray, batch_reach2::AbstractArray, batch_info1::AbstractArray, batch_info2::AbstractArray)

ModelVerification.has_two_reach_node — Method

has_two_reach_node(prop_method, model_info, node)

Checks whether there are two nodes connected to the current node, i.e., there are two previous nodes.

ModelVerification.is_activation — Method

is_activation(l)

ModelVerification.next_nodes — Method

next_nodes(prop_method::BackwardProp, model_info, node)

Returns the previous nodes of the node for BackwardProp methods. Since this is for BackwardProp methods, the previous nodes are the "next" nodes.

ModelVerification.next_nodes — Method

next_nodes(prop_method::ForwardProp,  model_info, node)

Returns the next nodes of the node for ForwardProp methods.

ModelVerification.output_node — Method

output_node(prop_method::BackwardProp, model_info)

Returns the starting nodes of the model for BackwardProp methods. Since this is for BackwardProp methods, the starting nodes of the model are the output nodes.

ModelVerification.output_node — Method

output_node(prop_method::ForwardProp, model_info)

Returns the final nodes of the model.

ModelVerification.prev_nodes — Method

prev_nodes(prop_method::BackwardProp, model_info, node)

Returns the next nodes of the node for BackwardProp methods. Since this is for BackwardProp methods, the next nodes are the "previous" nodes.

ModelVerification.prev_nodes — Method

prev_nodes(prop_method::ForwardProp,  model_info, node)

Returns the previous nodes of the node for ForwardProp methods.

ModelVerification.propagate — Method

propagate(prop_method::PropMethod, model_info, batch_info)

ModelVerification.propagate_act_batch — Method

propagate_act_batch(prop_method::ForwardProp, σ, batch_reach::AbstractArray, batch_info)

ModelVerification.propagate_layer_batch — Method

propagate_layer_batch(prop_method, layer, batch_bound, batch_info)

ModelVerification.propagate_layer_method — Method

propagate_layer_method(prop_method::BackwardProp, model_info, batch_info, node)

ModelVerification.propagate_layer_method — Method

propagate_layer_method(prop_method::ForwardProp, model_info, batch_info, node)

ModelVerification.propagate_linear_batch — Method

propagate_linear_batch(prop_method::ForwardProp, layer, batch_reach::AbstractArray, batch_info)

ModelVerification.propagate_skip_batch — Method

propagate_skip_batch(prop_method::ForwardProp, layer, batch_reach1::AbstractArray, batch_reach2::AbstractArray, batch_info)

ModelVerification.propagate_skip_method — Method

propagate_skip_method(prop_method::BackwardProp, model_info, batch_info, node)

ModelVerification.propagate_skip_method — Method

propagate_skip_method(prop_method::ForwardProp, model_info, batch_info, node)

NOTE: Need to include ForwardProp, BackwardProp, ... from solvers\solver.jl.

Bivariate

ModelVerification.propagate_skip — Method

propagate_skip(prop_method::AlphaCrown, layer::typeof(+), bound1::AlphaCrownBound, bound2::AlphaCrownBound, batch_info)

ModelVerification.propagate_skip — Method

propagate_skip(prop_method, layer::typeof(+), bound1::ImageStarBound, bound2::ImageStarBound, batch_info)

ModelVerification.propagate_skip — Method

propagate_skip(prop_method, layer::typeof(+), bound1::ImageZonoBound, bound2::ImageZonoBound, batch_info)

Convolution

ModelVerification.bound_layer — Method

bound_layer(layer::Conv{2, 4, typeof(identity), Array{Float32, 4}, Vector{Float32}}, lower_weight::AbstractArray, upper_weight::AbstractArray, lower_bias::AbstractArray, upper_bias::AbstractArray)

ModelVerification.bound_onside — Method

bound_onside(layer::Conv{2, 4, typeof(identity), Array{Float32, 4}, Vector{Float32}}, conv_input_size::AbstractArray, batch_reach::AbstractArray)

ModelVerification.interval_propagate — Function

interval_propagate(layer::Conv{2, 4, typeof(identity), Array{Float32, 4}, Vector{Float32}}, interval, C = nothing)

ModelVerification.propagate_by_small_batch — Method

propagate_by_small_batch(f, x; sm_batch=500)

ModelVerification.propagate_linear — Method

propagate_linear(prop_method::ImageStar, layer::Conv, bound::ImageStarBound, batch_info)

ModelVerification.propagate_linear — Method

propagate_linear(prop_method::ImageStar, layer::ConvTranspose, bound::ImageStarBound, batch_info)

ModelVerification.propagate_linear — Method

propagate_linear(prop_method::ImageZono, layer::Conv, bound::ImageZonoBound, batch_info)

ModelVerification.propagate_linear — Method

propagate_linear(prop_method::ImageZono, layer::ConvTranspose, bound::ImageZonoBound, batch_info)

Dense

Identity

Normalise

ReLU

Stateless